+++ example/simple.php
@@ -10,7 +10,19 @@ class AlertJob implements Krak\Job\Job
     public function handle() {}
 }
 
-$kernel = Krak\Job\createKernel(new Predis\Client());
-$dispatch = $kernel->createDispatch();
+$kernel = new Krak\Job\Kernel();
+$kernel->config([
+    'queue' => 'jobs',
+    'sleep' => 10,
+]);
+$kernel->queueManager(function() {
+    return Krak\Job\createQueueManager(new Predis\Client());
+});
 
-Krak\Job\registerConsole($app, $kernel);
+if ($argv[1] == 'dispatch') {
+    $kernel->dispatch(new AlertJob(1));
+} else {
+    $app = new Symfony\Component\Console\Application();
+    Krak\Job\registerConsole($app, $kernel);
+    $app->run();
+}