+++ src/Dispatch.php
@@ -6,6 +6,8 @@ interface Dispatch
 {
     /** @return WrappedJobBuilder */
     public function wrap(Job $job);
+    /** wraps and dispatches a job */
+    public function dispatch(Job $job);
     /** Dispatches a wrapped job */
-    public function dispatch(WrappedJob $job);
+    public function dispatchWrappedJob(WrappedJob $job);
 }