+++ test/functional/produce.php
@@ -9,6 +9,7 @@ require_once __DIR__ . '/../../vendor/autoload.php';
 class EchoJob implements Job\Job
 {
     private $id;
+
     public function __construct($id) {
         echo "$id\n";
         $this->id = $id;
@@ -20,7 +21,7 @@ class EchoJob implements Job\Job
 }
 
 $kernel = require __DIR__ . '/kernel.php';
-$dispatch = $kernel->createDispatch();
+$dispatch = $kernel['dispatch'];
 
 $num = intval($argv[2]);